+++ sdk/dataai/azure-data-ai/azure/data/ai/aio/_client.py
@@ -16,8 +16,8 @@ from azure.core.pipeline import policies
 from azure.core.rest import AsyncHttpResponse, HttpRequest
 
 from .._utils.serialization import Deserializer, Serializer
-from ._configuration import AzureDataAIClientConfiguration
-from ._operations import _AzureDataAIClientOperationsMixin
+from ._configuration import InferenceClientConfiguration
+from ._operations import _InferenceClientOperationsMixin
 
 if sys.version_info >= (3, 11):
     from typing import Self
@@ -28,9 +28,7 @@ if TYPE_CHECKING:
     from azure.core.credentials_async import AsyncTokenCredential
 
 
-class AzureDataAIClient(
-    _AzureDataAIClientOperationsMixin
-):  # pylint: disable=docstring-keyword-should-match-keyword-only
+class InferenceClient(_InferenceClientOperationsMixin):  # pylint: disable=docstring-keyword-should-match-keyword-only
     """Azure Data AI provides semantic reranking operations.
 
     :param endpoint: The Azure Data AI endpoint. Required.
@@ -50,7 +48,7 @@ class AzureDataAIClient(
         self, endpoint: str, credential: Union["AsyncTokenCredential", AzureKeyCredential], **kwargs: Any
     ) -> None:
         _endpoint = "{endpoint}"
-        self._config = AzureDataAIClientConfiguration(endpoint=endpoint, credential=credential, **kwargs)
+        self._config = InferenceClientConfiguration(endpoint=endpoint, credential=credential, **kwargs)
 
         _policies = kwargs.pop("policies", None)
         if _policies is None: